# parking-reserve-back/swagger.yaml
openapi: 3.0.0
info:
  title: Node.js Sign-up and Verification API
  description: Node.js + MongoDB - API with email sign-up, verification, authentication and forgot password
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /parking_slots:
    get:
      summary: Get a list of slots
      operationId: getAllParkingSlots
      responses:
        "200":
          description: An array of all accounts
          content:
            application/json:
              # schema:
              #   type: array
              #   items:
              #     type: object
              #     properties:
              #       id:
              #         type: string
              #         example: "5eb12e197e06a76ccdefc121"
              #       title:
              #         type: string
              #         example: "Mr"
              #       firstName:
              #         type: string
              #         example: "Jason"
              #       lastName:
              #         type: string
              #         example: "Watmore"
              #       email:
              #         type: string
              #         example: "jason@example.com"
              #       role:
              #         type: string
              #         example: "Admin"
              #       created:
              #         type: string
              #         example: "2020-05-05T09:12:57.848Z"
              #       updated:
              #         type: string
              #         example: "2020-05-08T03:11:21.553Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Create a new parking slot
      description: Restricted to admin users.
      operationId: createParkingSlot
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            # schema:
            #   type: object
            #   properties:
            #     title:
            #       type: string
            #       example: "Mr"
            #     firstName:
            #       type: string
            #       example: "Jason"
            #     lastName:
            #       type: string
            #       example: "Watmore"
            #     email:
            #       type: string
            #       example: "jason@example.com"
            #     password:
            #       type: string
            #       example: "pass123"
            #     confirmPassword:
            #       type: string
            #       example: "pass123"
            #     role:
            #       type: string
            #       enum: [Admin, User]
            #   required:
            #     - title
            #     - firstName
            #     - lastName
            #     - email
            #     - password
            #     - confirmPassword
            #     - role
      responses:
        "200":
          description: Parking slot created successfully.
          content:
            application/json:
            #   schema:
            #     type: object
            #     properties:
            #       id:
            #         type: string
            #         example: "5eb12e197e06a76ccdefc121"
            #       title:
            #         type: string
            #         example: "Mr"
            #       firstName:
            #         type: string
            #         example: "Jason"
            #       lastName:
            #         type: string
            #         example: "Watmore"
            #       email:
            #         type: string
            #         example: "jason@example.com"
            #       role:
            #         type: string
            #         example: "Admin"
            #       created:
            #         type: string
            #         example: "2020-05-05T09:12:57.848Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /parking_slots/{id}:
    parameters:
      - in: path
        name: id
        description: Account id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    get:
      summary: Get a single account by id
      description: Admin users can access any account, regular users are restricted to their own account.
      operationId: getAccountById
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Details of the specified account
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  title:
                    type: string
                    example: "Mr"
                  firstName:
                    type: string
                    example: "Jason"
                  lastName:
                    type: string
                    example: "Watmore"
                  email:
                    type: string
                    example: "jason@example.com"
                  role:
                    type: string
                    example: "Admin"
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  updated:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    put:
      summary: Update an account
      description: Admin users can update any account including role, regular users are restricted to their own account and cannot update role.
      operationId: updateAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Mr"
                firstName:
                  type: string
                  example: "Jason"
                lastName:
                  type: string
                  example: "Watmore"
                email:
                  type: string
                  example: "jason@example.com"
                password:
                  type: string
                  example: "pass123"
                confirmPassword:
                  type: string
                  example: "pass123"
                role:
                  type: string
                  enum: [Admin, User]
      responses:
        "200":
          description: Account updated successfully. The details of the updated account are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  title:
                    type: string
                    example: "Mr"
                  firstName:
                    type: string
                    example: "Jason"
                  lastName:
                    type: string
                    example: "Watmore"
                  email:
                    type: string
                    example: "jason@example.com"
                  role:
                    type: string
                    example: "Admin"
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  updated:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"